Extracted PeerOskDomainSeperator
